Add dedicated errors file
