+++ packages/cli/src/lib/index.ts
@@ -3,5 +3,6 @@ export * from './validation';
 export * from './compile';
 export * from './config';
 export * from './diff';
+export * from './errors';
 export * from './files';
 export * from './transform';
